+++ src/context/product.context.jsx
@@ -1,0 +1,16 @@
+import React, { createContext, useState } from "react";
+import DB from "../data.json";
+
+export const ProductContext = createContext({
+  products: [],
+});
+
+export const ProductProvider = ({ children }) => {
+  const [products, setProducts] = useState(DB);
+
+  return (
+    <ProductContext.Provider value={{ products, setProducts }}>
+      {children}
+    </ProductContext.Provider>
+  );
+};